requirements.txt
flask
gunicorn
quant-platform-kit @ git+https://github.com/QuantStrategyLab/QuantPlatformKit.git@v0.7.7
us-equity-strategies @ git+https://github.com/QuantStrategyLab/UsEquityStrategies.git@v0.7.8
pandas
requests
pandas_market_calendars
pytz
longport
schwab-py
google-cloud-secret-manager
google-cloud-storage
google-auth
numpy
